added filteredgrid.hh
A very basic implementation of a FilteredGridPart, that will work only with
LeafGridParts and HierarchicGridParts and that doesn't offer
iterators for codim \neq 0.

The intersection iterator has to be tested!

includes example implementation of a Filter that chooses Elements in a
sphere of given radius and center.

how to use a FilteredGridPart with RadialFilter:

  typedef HierarchicGridPart<GridType> GridPartType;  // the underlying GridPart
  typedef RadialFilter<GridType> FilterType;   // type of the filter we use
  Dune::FieldVector<GridType::ctype,GridType::dimension> C(0.5);
  FilterType filter(C, 0.5);
  typedef FilteredGridPart<GridPartType,FilterType> FGridPartType;
  FGridPartType part(grid,filter);


git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/trunk@948 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0