EntityFunction interface class for functions, which provide an evaluation method in a quadrature-style on entities, e.g.
  f(entity, quad, qP, ret)

Motivation is that there are functions, which are neither pure analytical
 (so cannot be evaluated as a Dune::Function) nor DiscreteFunctions.
For example any nontrivial f(df), where df is a DiscreteFunction, but f some
analytical function operating on the values of df.

A simple wrapper EntityFunctionWrapper<FunctionType> translates a
Dune::Function into a EntityFunction

These functions are currently used in the LagrangeInterpolator and the
fuelcell potential subproblem.


git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/trunk@989 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0