corrected unitRow => prevent producing lots of zero entries
added range-checks in some access methods

added consistencyCheck method, as many methods currently result in a
inconsistency between the counters for nonzeros and the really stored elements.
The method is called as argument of an assert() 
at beginning and at end of important non-const methods.
By this, buggy methods can be identified. 
The check is only performed if the variable checkNonConstMethods is set, which 
if off by default.



git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/trunk@1249 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0