*Important Change*: GridParts now deliver iterators for all partition types
To wrap a GridPart into a GridView (and then use it, e.g., in the VTKWriter),
the GridPart must be able to change the iterator type. Additionally, the pitype
template argument has been removed from the standard GridParts. The default
iterator (returned by begin< 0 >()) remains an InteriorBorder_Partition iterator
(which was the default pitype anyway).

Moreover, GridParts now have an indexSetPartitionType, which is the largest
partition type, for which the index set provides indices.


git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/trunk@4015 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0