This adds an interface to PETSc to dune-fem. As of now, there is a PetscDiscreteFunction, a PetscLinearOperator and a PetscSolver. You can see them working in a 'real world example' in
dune/fem/petsc/tests/l2projection.

See also the README file there. A very short not on how to install
PETSc (at least a configuration that is working for me on my machine)
can be found in 

dune/fem/petsc/README.

If you want to see the history of this project up to this point, check
out the dune-fem-petsc module from the svn servers at Freiburg. From
now on, the development of the PETSc interface to dune-fem will
be done here.


git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/trunk@7404 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0