do not store the index state in the persistent container
Previously, the persistent container was not allowed to delete entries
on resize. This is due to the fact that, apart from the index, its use
state was stored in the persistent container. Storing it in a separate
vector allows us to check the index state without a remaining reference
in the persistent container itself.