[bugfix] do not store reference to a temporary
In DefaultBasisFunctionSet::jacobianAll/hessianAll, a reference to the geometry
was stored in the transformation functor. This was a reference to a temporary
going out of scope directly after the creation of the functor. The patch
fixes this issue by increasing the lifetime of the geometry object to
the lifetime of the function.

Kudos for this fix go to Sven Kaulmann <sven.kaulmann@uni-muenster.de>.