[Interpolation] Added interpolation method to DG spaces
Some discrete function spaces provide a projection from a given local
function onto the local discrete function space by the public member
method:

  void interpolate ( const LocalFunction &lf, LocalDofVector &dofs ) const;

Calling this method several times on a fixed entity may be inefficient,
e. g. if a local mass matrix is assembled and inverted on a each
call. This patch introduces a new method:

  InterpolationType interpolation ( const EntityType &entity ) const;

which allows for storing the interpolation object.