[GridPartView] Adapt Dune::Fem::GridPartView to recent changes to Dune::GridView
GridViews in Dune now always model the All_Partition. The default
iterator of a grid view thus has to be of All_Partition type. In
addition this patch adds the missing methods

  int overlapSize ( int codim ) const;
  int ghostSize ( int codim ) const;

These methods cannot be implemented through a GridPart at the time.
Thus, the current implementation simply throws an Dune::NotImplemented
exception.