[bugfix] avoid duplicating the twistIn* methods in TwistUtility for GeometryGrid
The type of LeafIntersectionIterator and LevelIntersectionIterator of a
DUNE grid might equal (e.g., for AlbertaGrid or SPGrid). In this case,
they will also equal for a GeometryGrid over them. While we may still
export both typedefs, we cannot have the methods twistIn* for both,
LeafIntersection and LevelIntersection.

To resolve the issue, we simply provide a template method taking any
intersection and then call the TwistUtility of the host grid with the
host intersection.

The main difference is that we can no longer obtain the type of the host
intersection (as we do not know whether we have a leaf of level
intersection). On the upside, why do we care?

Apart from fixing the bug, this code is simpler and more readable.
However, due to reduced type safety, it might make error messages harder
to understand.