[localfunction] add LocalFunctionConverter
This LocalFunctionConverter converts a give localGridFunction into a
localGridFunction on a different FunctionSpace.
The restriction/prolongation converter needs to be specified by the user.
It basically stores a reference/copy to the hostLocalFunction, evaluate it in its
own RangeType and converts this RangeType into the new RangeType.

This class is designed to obtain a certain sub component of the gridfunction,
eg. the density, velocity or engery component for the Euler system if the gridfunction
returns a tuple ( \rho, \rho u, E ).