[AdaptiveLeafIndexSet] Correctly convert grid part to grid entities and vice versa
The implementation of Dune::Fem::AdaptiveLeafIndexSet builds on
Dune::Fem::CodimIndexSet which expects a Dune::Grid as template
argument; all arguments passed to members of Dune::Fem::CodimIndexSet
must be compatible with the template grid type. If entities returned by
the grid and a grid part respectively differ in type (as is the
case in some meta grid parts), type conversions as provided by
Dune::Fem::gridEntity() and Dune::Fem::GridPart::convert() must be
called explicitly. This patch adds the missing calls to these methods in
Dune::Fem::AdaptiveLeafIndexSetBase.