make adaptation example work (in serial)
In principle, we virtualize the restrict-prolong object. A discrete then
implicitly casts into the default restrict-prolong object, so we can pass
a discrete function to adapt and loadBalance.

For the parallel version, we have to enable the methods addToList and
removeFromList on the VirtualizedRestrictProlong object again. Unfortunately,
these cause a segfault even in serial runs (double-free exception?).