add rewrite of form compiler
The original form compiler does many things at once and is quite hard to
understand and extend.

This rewrite is now split up:
- the form can be written without knowing generator or even model
- a function translates a form equation into an elliptic model
- the code is written out by a separate function

At the core, the implementation now uses the ufl translator to traverse the
expression tree. This makes the actual transformations more readable (though
documentation will be necessary, here).

In the current state, the model cannot be compiled into a Python module and
the many possible nodes in the tree are not parsed, yet. It is just a first step.
A first glimpse at the goal can be taken in 'demo/ellipticmodel.py'. A lot of
convenience is still missing, though.