[bugfix] need to decrement base iterator before erasing
The std::reverse_iterator::base will actually point to the object following the
one the reverse_iterator points to. To delete the correct object, we have to
decrement the base iterator before passing it to erase.

This bug caused segmentation faults, when the most recently inserted function,
i.e., the last in the list, is removed again.