make dune-fem adaptation work again
Note: Instead of calling
          hgrid.adapt([phi])
      you now call
          dune.fem.adapt(hgrid, [phi])

Compiling the grid adaptation into the grid module mixes parts of FemPy
and CorePy. Therefore, the adaptation process is now compiled into a
separate module (called adapt_<hash>).

Instead of manipulating the deleter of the hierarchical grid, we now
let Python handle the deletion of the grid adaptation object via a
weak reference. This is the same mechanism we also use for a grid parts.