improve exporting of spaces (back) to Python
Discrete functions and schemes (referred to as parent below) can return the
stored space to the Python user.
As this is a reference, we normally have to keep the parent alive until the
returned space is garbage collected. However, frequently the space was passed
in from Python itself. In this case, we directly return the space's Python
object, so we don't artificially need to keep the parent alive. Otherwise, we
return the guarded reference as before.

Bug: For some reason, PyBind11 seems to require a move constructor to pass a
reference into Python. For this reason, passing out references is disabled for
now and only spaces originating from Python can be returned.