[bugfix] directly implement UniqueTopology
During the removal of the genericgeometry subdirectory, the
GenericGeometry::Topology was removed, too. Unfortunately, it was used in
the orthonormal shape function set to compute a "unique" topology type.
This patch directly implements the construction of this type by a helper
template called "Dune::Fem::Impl::UniqueTopology".

Apart from fixing the bug, this reimplementation makes the code more readable.