simplify IndexSetDofMapper::onSubEntity
Currently, the method IndexSetDofMapper::onSubEntity is implemented by a copy
of the MapFunctor. It thus inherits the reordering of degrees of freedom to
eliminate the twist problem for Lagrange spaces.

However, the question whether a local DoF is assigned to a subentity or not does
not depend on the ordering of the global indices. Eliminating the corresponding
checks heavily simplifies the functor (and eliminates to std::vectors).

This patch goes one stop further and integrates the functor into a lambda, thus
keeping the implementation within the method body.