[dgspace] add TupleDiscontinuousGalerkinSpace
A TupleDiscontinuousGalerkinSpace create the product space of given
DiscontinuousGalerkinSpaces. In contrast to a TupleDiscreteFunctionSpace,
the product space is constructed on the level of shape function sets.
As a result, the DoF mapper remains a CodimensionMapper and the localBlockSize
is the sum of the underlying localBlockSizes.

In the context of pure DG computations, this space is slightly more efficient
than the TupleDiscreteFunctionSpace.