[ci] disable debian:8-backports--clang image due to bugs in gcc-4.9 standard library
This image uses clang in C++14 mode together with the standard library
from gcc-4.9. Unfortunately, this library version marks the methods
imag() and real() constexpr but not const. In C++14, constexpr no longer
implies const, so that no const version is available. This causes our
build to fail.

As this is not a dune-fem bug, we simply disable this image for our
tests.