Rename variable to "correct" names
Rationale: the instance of a DofConstraints class models certain
constraints, hence it should be named "constraints" or similar.

The solution operator-method inside the class DofConstraints has been
renamed from operator() to coinstrain. Hence in order to constrain  a
given DoF-Vector to the prescribed values, the call "should" be (nameing
proposol)

ConstraintType constraints(all needed constructor arguments)

constraints.constrain(victimDF);

Note the the DofConstraints class implements as well kind of an operator
interaface as well as an inverse-operator which installs the prescribed
constraints into a given discrete function.