allow implicit conversion from `std::initializer_list` to `ExplicitFieldVector`
This constructor is similar to the conversion from anther `DenseVector`,
i.e., all elements of the `std::initializer_list` are ensured to be of type
`T`.

Having this constructor implicit is useful when initializing a
`std::tuple` containing different range types from initializer lists.
This is, for example, used in dune-fempy.