Allow storage of a BindableGridFunction by reference if explicitly requested.
The usage case is as follows: one stores a BindableGridFunction in,
e.g., a model class as a copy. Then the BGF is wrapped in to a
ConstLocalFunction in order to add the CLF interface. This makes a
second copy. However, this second copy is completely useless as the
model already holds a copy and is also a "bindable object". In this case
we only need the LF-interface methods, but not the additional storage.