Change behavior of multithreaded section of code: use availbale number of threads by default instead of running in single thread mode.
Introduced new environment variable to set number of threads: `DUNE_NUM_THREADS`
(this is in addition to existing `OMP_NUM_THREADS` which was also used for pthreads).

Also bugfix in `BindableGF` which fixes an issue with `GeometryGP` and multithreading.

additional assert only meaningful in parallel mode

add 'initMultThreadMode' to OMP thread manager