started on exporting the fem gridpart directly without using the GridView wrapper when exporting to Python.
Started with filtered and adaptive gp.

Added an test example for the filtered gp (taken directly from tutorial).

The idea is to use the generator for the gv's from dune-grid. That still needs a small bit of work
but should work now in principle. Important is that the 'registerGridView' from dune/fempy/py/gridview.hh
is taken so that 'canadapt' is added and the 'mark' methods to the adaptive gp.

Since all methods on the gv are exported this also shows incompatibilities between gp's and gv's.
For exaple the filtered gp is missing a 'size(codim)' method (not yet fixed).