[parallel,lagrange] assemble matrix on AllPartition, but residual only in interior
Currently operator and residual are assembled only in the interior. This
breaks parallel computations with conforming elements. At the same time
any operator evaluation (e.g. computing the residual) ends with a global
sum, so that updates on different ranks are summed up.

There are different possible approaches to fix this.

 1. decomposition of DOFs
   - assign each DOF to a unique rank (e.g. lowest rank number)
   - assemble matrix & residual on AllPartition
   - residual is automatically consistent, no communication required
   - norms are only evaluated for DOFs owned by the rank

 2. decomposition of the
   - assemble matrix on AllPartition, operators only in interior
   - residual needs to be communicated (all-to-all) and summed up to  have the complete residual.
   - as operators are evaluated only in the interior, the norm is automatically consistent