linearized scheme, fix some trivial things and change setup().
Obviously, linerarized has not been used for a long time and hence did
not longer work (so may be generate a test for that if I find the time).

The commit also changes setup(fct) to allow mere grid-functions and
introduces setup(<nothing>) to implicitly linearize around zero.

The latter is meant for the case I am interested, namely to "linearize"
affine linear operators for the sake of having a pre-assembled scheme +
solve method.