Schemes, add a third parameter to solve(rhs, target, additiveConstraints)
The purpose is to add any present constraints already present in rhs to
the rhs of the system to solve. The "normal" behaviour in Fem is to
apply zero constraints to rhs in order to avoid that the Dirichlet
values from the model are spoiled by spurious integration artifacts
generated by boundary basis functions.