Schemes linearized + Python: export the pure linear operator (without shift)
Why?

The use case is a Uzawa algorithm where I would like to allow RHS
contributions to be contained in the principal part A

A | B   u   f
----- * - = -
Bt| 0   p   g (often just 0)

Now, as the Uzawa in principle is just a variant of the (P)CG-method for
Schur's complement, all RHS contributions are handled by adding them to
the initial residual in the setup step.

During the iteration, one really just needs the linear part of A.

It is difficult to work around this because Schur's complement requires
the inversion of A. Here linearized(scheme) works nicely in the setup
script, but its solve method always takes any affine shift into account
and this spoils the subsequent inversions in the iteration "phase" of
the Uzawa algo.

Also: the other schemes also have a linear() method.